Available Models

Available Models

Recurrent models with hand-coded gradients

RNNFluxes.LSTMModel — Type.

type LSTMModel

Implementation of an LSTM with handcoded gradients. It has the following constructor:

LSTMModel(nVar,nHid; dist = Uniform, forgetBias = 1, nDropout=nHid ÷ 10)

Parameters

nVar number of input (predictor) variables for each time step

nHid number of hidden nodes contained in the LSTM

dist distribution to generate initial Weights, currently Uniform and Normal are supported, defaults to Uniform

forgetBias determine bias of generating forget bias weights, defaults to 1

nDropOut number of dropout nodes for each prediction, defaults to 10% of the nodes

source

RNNFluxes.RNNModel — Type.

type RNN

Implementation of an RNN with handcoded gradients.

RNNModel(nVar,nHid;dist=Uniform)

Parameters

nVar number of input (predictor) variables for each time step

nHid number of hidden nodes contained in the RNN

dist distribution to generate initial Weights, currently Uniform and Normal are supported, defaults to Uniform

source

Trainning a generic model using AutoDiff

We have also wrapped the training and plotting machinery here together with the ForwardDiff package, so that you can invert any custom model, too. Please note that, for this prupose the package (Flux.jl)[https://github.com/FluxML/Flux.jl] might be a better and well-maintained option.

RNNFluxes.GenericModel — Type.

type GenericModel

Type wrapping a generic model that is differentiated with AutoGrad. Only a predict function is needed.

source

A short example on how to define and train such a generic model follows. Let's assume we already know the model structure but want to estimate the parameters. Any Julia code would be valid here in defining this model:

function my_generic_model{T}(model,w::AbstractArray{T},x)

    a = w[1]
    b = w[2]
    c = w[3]

    map(x) do ix
        cumsum(a * ix[1,:])+exp.(b * ix[2,:])+cumprod(c * ix[3,:])
    end
end

We generate some artificial data again:

nTime=100
nSample=200
nVarX = 3
x = [rand(nVarX,nTime) for i=1:nSample]

true_model(x) = transpose(cumsum(x[1,:])+exp.(x[2,:])+cumprod(1.7*x[3,:]))

y = true_model.(x);

In the next step we define a GenericModel which includes our above-mentioned predict function.

params = ()
w      = rand(3) #Init weights
m      = RNNFluxes.GenericModel(params,w,my_ANN_model)

Training works "as usual"

t = train_net(m,x,y,4001);