Training and predicting models

Training and predicting models

RNNFluxes.train_net — Function.

train_net(model::FluxModel, x, y, nEpoch)

Trains a model defined in model on the given predictors x and labels y for nEpoch training steps. Here, x as well as y are vectors of matrices of size (nVar x nTimeStep). Each matrix stands for a single sample, for example a tower time series. The length of the individual time series is allowed to differ.

Keyword arguments

lossFunc lossFUnction to be evaluated, defaults to mseLoss

batchSize number of samples to be used for each update step

searchParams One of the training methods defined in Knet. Can be any object of type Sgd, Momentum, Adam, Adagrad, Adadelta, Rmsprop. Have a look at update.jl to see how to construct this

infoStepSize stepsize in which traing and validation losses will be calculated and printed to the screen

valiFrac fraction of samples to be omitted from training as a validation dataset. Can be either a Float for a random fraction or vector of indices of omitted datasets, defaults to 0.2

losscalcsize determines how often the losses of the whole dataset will be evaluated so that the training process can be analyzed afterwards

plotProgress shall the training be visualized, works only in Jupyter notebooks

nPlotsample number of sample points used for plotting

source

RNNFluxes.predict_after_train — Function.

Given the trained model this function does predictions on another data set it also does the back-transform of the predicted values from (0,1) to original range whole thing should be improved. An object of Type TrainedModel should be used or sth like a list in R

source